Allow other elements to start the gallery
Allow other elements to start Lightbox on behalf of other elements.
It allows them to act as a sort of <label> for another element that
actually contains the data-lightbox attribute.

The way it works is: you specify the data-lightbox-label-for
attribute for an element. The value of that attribute is the HTML
id of a link that contains a data-lightbox attribute of a rel
attribute containing "lightbox".